changed Javascript of printQRs, to leave the tab opened in chrome
Putting window.close directly after print causes the window to disappear before the print dialogue on chrome, because Chrome does not use system print dialogue. However, using setTimeout causes the close to be queued on the event queue, the 500ms timeout isn't actually a timing thing, the timer will be blocked until the print dialog is closed at which time the 500ms timer is started.