hide button after course creation
after course creation only one of the buttons to course overview should be shown to the user.
hiding the button was not enough as it still took its place in the DOM including in the layout.
so we needed to remove the naming for the link to make it totally invisible.