BC-2149 - new task: keep textarea in sync with ckeditor (#2943)
Fixing an issue where after creating a new task, assigning it to a course and adding a description, the description gets lost after clicking the save-button for the first time. The description of the the new task can be edited through a CKEditor-instance that wraps a hidden textarea. The fix takes care that both stay in sync.

Co-authored-by: Oliver Happe <OliverHappe@users.noreply.github.com>
Co-authored-by: Thomas Feldtkeller <thomas.feldtkeller@dataport.de>