BC-6735 Style all scrollbars (#3614)
The polyfills in base.js were removed because of a problem that arose with the update of caniuse-lite, which was part of the necessary update of autoprefixer. After this update the client didn't work anymore on older iOS versions (e.g. 15.8.3), though according to our browserslist spec it should have. It turned out that babel somehow did not transpile correctly when the condition `if (window.EventTarget) {...}` was present. Removing it solved the problem. The other two polyfills were removed together with it, because they are also unnecessary by now.